requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0
nltk==3.8.1
python-dotenv==1.0.0
requests==2.28.2
schedule==1.2.0
